+++ Changes.md
@@ -25,6 +25,7 @@ Breaking Changes
 ----------------
 
 - Gadget : Added new virtual method and private member variables.
+- ImageGadget : Removed/moved private member variables.
 
 0.58.3.0 (relative to 0.58.2.0)
 ========